Rejigger the schema a little
Instead of a rom having a (pathless) file to house its checksums and
size, let's just put the metadata in the rom table/object itself.
It makes the join syntax a bit nicer to read, so we don't have to
join the files table with itself to find checksum matches.